+++ src/behaviors/ProductTypeLayoutBehavior.php
@@ -5,31 +5,18 @@ use Ryssbowh\CraftThemes\Themes;
 use Ryssbowh\CraftThemes\interfaces\LayoutInterface;
 use Ryssbowh\CraftThemes\interfaces\ThemeInterface;
 use Ryssbowh\CraftThemes\services\LayoutService;
-use craft\base\Volume;
-use craft\elements\GlobalSet;
-use craft\models\CategoryGroup;
-use craft\models\EntryType;
-use craft\models\TagGroup;
+use craft\commerce\models\ProductType;
 use yii\base\Behavior;
 
 /**
- * This behaviour is attached to some Craft elements (CategoryGroup, EntryType, Volume, TagGroup, GlobalSet)
- * 
- * It gives a shorter way to access those element's layouts, and edit displays/blocks urls.
- * $group = \Craft::$app->categories->getGroupById(1);
- * $layout = $group->getLayout('theme-handle');
+ * This behaviour is attached to all the variants
  */
-class LayoutBehavior extends Behavior
+class ProductTypeLayoutBehavior extends Behavior
 {
     /**
-     * @var CategoryGroup|EntryType|Volume|TagGroup|GlobalSet
+     * @var ProductType
      */
     public $owner;
-
-    /**
-     * @var string
-     */
-    public $type;
 
     /**
      * Layout getter
@@ -42,8 +29,21 @@ class LayoutBehavior extends Behavior
         if (is_null($theme)) {
             $theme = Themes::$plugin->registry->current;
         }
-        $uid = $this->type == LayoutService::USER_HANDLE ? '' : $this->owner->uid;
-        return Themes::$plugin->layouts->get($theme, $this->type, $uid);
+        return Themes::$plugin->layouts->get($theme, LayoutService::PRODUCT_HANDLE, $this->owner->uid);
+    }
+
+    /**
+     * Variant layout getter
+     * 
+     * @param  string|ThemeInterface|null $theme theme instance or theme handle, or null for the current theme
+     * @return ?LayoutInterface
+     */
+    public function getVariantLayout($theme = null): ?LayoutInterface
+    {
+        if (is_null($theme)) {
+            $theme = Themes::$plugin->registry->current;
+        }
+        return Themes::$plugin->layouts->get($theme, LayoutService::VARIANT_HANDLE, $this->owner->uid);
     }
 
     /**
@@ -69,5 +69,18 @@ class LayoutBehavior extends Behavior
     public function getEditBlocksUrl($theme = null): string
     {
         return $this->getLayout($theme)->getEditBlocksUrl();
+    }
+
+    /**
+     * Get the url to edit displays for a theme and a view mode.
+     * If the theme is null the current theme will be used.
+     * 
+     * @param  ThemeInterface|string|null    $theme
+     * @param  ViewModeInterface|string|null $viewMode
+     * @return ?string
+     */
+    public function getEditVariantDisplaysUrl($theme = null, $viewMode = null): ?string
+    {
+        return $this->getVariantLayout($theme)->getEditDisplaysUrl($viewMode);
     }
 }