+++ src/events/RegisterFieldDisplayerEvent.php
@@ -3,6 +3,7 @@ namespace Ryssbowh\CraftThemes\events;
 
 use Ryssbowh\CraftThemes\exceptions\FieldDisplayerException;
 use Ryssbowh\CraftThemes\interfaces\FieldDisplayerInterface;
+use Ryssbowh\CraftThemes\models\fieldDisplayers\AllowedQtyDefault;
 use Ryssbowh\CraftThemes\models\fieldDisplayers\AssetRenderFile;
 use Ryssbowh\CraftThemes\models\fieldDisplayers\AssetRendered;
 use Ryssbowh\CraftThemes\models\fieldDisplayers\AssetSlick;
@@ -13,6 +14,7 @@ use Ryssbowh\CraftThemes\models\fieldDisplayers\CheckboxesLabel;
 use Ryssbowh\CraftThemes\models\fieldDisplayers\ColourDefault;
 use Ryssbowh\CraftThemes\models\fieldDisplayers\Date;
 use Ryssbowh\CraftThemes\models\fieldDisplayers\DateTime;
+use Ryssbowh\CraftThemes\models\fieldDisplayers\DimensionsDefault;
 use Ryssbowh\CraftThemes\models\fieldDisplayers\DropdownLabel;
 use Ryssbowh\CraftThemes\models\fieldDisplayers\ElementLink;
 use Ryssbowh\CraftThemes\models\fieldDisplayers\ElementLinks;
@@ -27,9 +29,13 @@ use Ryssbowh\CraftThemes\models\fieldDisplayers\MultiSelectLabel;
 use Ryssbowh\CraftThemes\models\fieldDisplayers\NumberDefault;
 use Ryssbowh\CraftThemes\models\fieldDisplayers\PlainTextPlain;
 use Ryssbowh\CraftThemes\models\fieldDisplayers\PlainTextTruncated;
+use Ryssbowh\CraftThemes\models\fieldDisplayers\PriceDefault;
+use Ryssbowh\CraftThemes\models\fieldDisplayers\ProductRendered;
+use Ryssbowh\CraftThemes\models\fieldDisplayers\ProductVariantsRendered;
 use Ryssbowh\CraftThemes\models\fieldDisplayers\RadioButtonsLabel;
 use Ryssbowh\CraftThemes\models\fieldDisplayers\RedactorFull;
 use Ryssbowh\CraftThemes\models\fieldDisplayers\RedactorTruncated;
+use Ryssbowh\CraftThemes\models\fieldDisplayers\StockDefault;
 use Ryssbowh\CraftThemes\models\fieldDisplayers\TableDefault;
 use Ryssbowh\CraftThemes\models\fieldDisplayers\TagLabel;
 use Ryssbowh\CraftThemes\models\fieldDisplayers\TagRendered;
@@ -42,6 +48,8 @@ use Ryssbowh\CraftThemes\models\fieldDisplayers\UrlLink;
 use Ryssbowh\CraftThemes\models\fieldDisplayers\UserDefault;
 use Ryssbowh\CraftThemes\models\fieldDisplayers\UserRendered;
 use Ryssbowh\CraftThemes\models\fieldDisplayers\UserSlick;
+use Ryssbowh\CraftThemes\models\fieldDisplayers\VariantRendered;
+use Ryssbowh\CraftThemes\models\fieldDisplayers\WeightDefault;
 use yii\base\Event;
 
 class RegisterFieldDisplayerEvent extends Event
@@ -59,6 +67,7 @@ class RegisterFieldDisplayerEvent extends Event
     {
         parent::init();
         $this->registerMany([
+            AllowedQtyDefault::class,
             AssetSlick::class,
             AssetRendered::class,
             AssetRenderFile::class,
@@ -69,6 +78,7 @@ class RegisterFieldDisplayerEvent extends Event
             ColourDefault::class,
             Date::class,
             DateTime::class,
+            DimensionsDefault::class,
             DropdownLabel::class,
             EmailEmail::class,
             ElementLink::class,
@@ -83,9 +93,12 @@ class RegisterFieldDisplayerEvent extends Event
             NumberDefault::class,
             PlainTextPlain::class,
             PlainTextTruncated::class,
+            PriceDefault::class,
+            ProductRendered::class,
             RadioButtonsLabel::class,
             RedactorFull::class,
             RedactorTruncated::class,
+            StockDefault::class,
             TableDefault::class,
             TagLabel::class,
             TagTitle::class,
@@ -98,6 +111,9 @@ class RegisterFieldDisplayerEvent extends Event
             UserDefault::class,
             UserRendered::class,
             UserSlick::class,
+            ProductVariantsRendered::class,
+            VariantRendered::class,
+            WeightDefault::class
         ]);
     }