+++ src/events/RegisterFieldsEvent.php
@@ -2,15 +2,21 @@
 namespace Ryssbowh\CraftThemes\events;
 
 use Ryssbowh\CraftThemes\exceptions\FieldException;
+use Ryssbowh\CraftThemes\models\fields\AllowedQty;
 use Ryssbowh\CraftThemes\models\fields\Author;
 use Ryssbowh\CraftThemes\models\fields\CraftField;
 use Ryssbowh\CraftThemes\models\fields\DateCreated;
 use Ryssbowh\CraftThemes\models\fields\DateUpdated;
+use Ryssbowh\CraftThemes\models\fields\Dimensions;
 use Ryssbowh\CraftThemes\models\fields\ElementUrl;
+use Ryssbowh\CraftThemes\models\fields\ExpiryDate;
 use Ryssbowh\CraftThemes\models\fields\File;
 use Ryssbowh\CraftThemes\models\fields\Matrix;
 use Ryssbowh\CraftThemes\models\fields\MatrixField;
 use Ryssbowh\CraftThemes\models\fields\PostDate;
+use Ryssbowh\CraftThemes\models\fields\Price;
+use Ryssbowh\CraftThemes\models\fields\Sku;
+use Ryssbowh\CraftThemes\models\fields\Stock;
 use Ryssbowh\CraftThemes\models\fields\Table;
 use Ryssbowh\CraftThemes\models\fields\TableField;
 use Ryssbowh\CraftThemes\models\fields\TagTitle;
@@ -21,6 +27,8 @@ use Ryssbowh\CraftThemes\models\fields\UserLastLoginDate;
 use Ryssbowh\CraftThemes\models\fields\UserLastName;
 use Ryssbowh\CraftThemes\models\fields\UserPhoto;
 use Ryssbowh\CraftThemes\models\fields\UserUsername;
+use Ryssbowh\CraftThemes\models\fields\Variants;
+use Ryssbowh\CraftThemes\models\fields\Weight;
 use yii\base\Event;
 
 class RegisterFieldsEvent extends Event
@@ -55,6 +63,14 @@ class RegisterFieldsEvent extends Event
         $this->register(UserPhoto::class);
         $this->register(UserEmail::class);
         $this->register(ElementUrl::class);
+        $this->register(ExpiryDate::class);
+        $this->register(Variants::class);
+        $this->register(Stock::class);
+        $this->register(Sku::class);
+        $this->register(Price::class);
+        $this->register(Dimensions::class);
+        $this->register(Weight::class);
+        $this->register(AllowedQty::class);
     }
 
     /**