+++ src/models/fieldDisplayers/PlainTextPlain.php
@@ -3,6 +3,7 @@ namespace Ryssbowh\CraftThemes\models\fieldDisplayers;
 
 use Ryssbowh\CraftThemes\models\FieldDisplayer;
 use Ryssbowh\CraftThemes\models\fieldDisplayerOptions\PlainTextFullOptions;
+use Ryssbowh\CraftThemes\models\fields\Sku;
 use Ryssbowh\CraftThemes\models\fields\UserFirstName;
 use Ryssbowh\CraftThemes\models\fields\UserLastName;
 use Ryssbowh\CraftThemes\models\fields\UserUsername;
@@ -39,7 +40,7 @@ class PlainTextPlain extends FieldDisplayer
      */
     public static function getFieldTargets(): array
     {
-        return [PlainText::class, UserFirstName::class, UserLastName::class, UserUsername::class];
+        return [PlainText::class, UserFirstName::class, UserLastName::class, UserUsername::class, Sku::class];
     }
 
     /**