+++ src/models/fields/ExpiryDate.php
@@ -8,21 +8,21 @@ use Ryssbowh\CraftThemes\models\layouts\EntryLayout;
 use Ryssbowh\CraftThemes\models\layouts\ProductLayout;
 
 /**
- * Handles the postDate value of entry elements
+ * Handles the dateCreated value of elements
  */
-class PostDate extends Field
+class ExpiryDate extends Field
 {
     /**
      * @var boolean
      */
     public $hidden = true;
-    
+
     /**
      * @inheritDoc
      */
     public static function getType(): string
     {
-        return 'post-date';
+        return 'expiry-date';
     }
 
     /**
@@ -30,7 +30,13 @@ class PostDate extends Field
      */
     public static function shouldExistOnLayout(LayoutInterface $layout): bool
     {
-        return $layout instanceof EntryLayout or $layout instanceof ProductLayout;
+        if ($layout instanceof ProductLayout) {
+            return true;
+        }
+        if ($layout instanceof EntryLayout) {
+            return ($layout->element->section->type != 'single');
+        }
+        return false;
     }
 
     /**
@@ -38,7 +44,7 @@ class PostDate extends Field
      */
     public function getHandle(): string
     {
-        return 'postDate';
+        return 'expiryDate';
     }
 
     /**
@@ -46,6 +52,6 @@ class PostDate extends Field
      */
     public function getName(): string
     {
-        return \Craft::t('themes', 'Date posted');
+        return \Craft::t('themes', 'Expiry date');
     }
 }