+++ src/models/fields/Price.php
@@ -4,25 +4,19 @@ namespace Ryssbowh\CraftThemes\models\fields;
 use Ryssbowh\CraftThemes\Themes;
 use Ryssbowh\CraftThemes\interfaces\LayoutInterface;
 use Ryssbowh\CraftThemes\models\Field;
-use Ryssbowh\CraftThemes\models\layouts\EntryLayout;
-use Ryssbowh\CraftThemes\models\layouts\ProductLayout;
+use Ryssbowh\CraftThemes\models\layouts\VariantLayout;
 
 /**
- * Handles the postDate value of entry elements
+ * Handles the price of variants or products
  */
-class PostDate extends Field
-{
-    /**
-     * @var boolean
-     */
-    public $hidden = true;
-    
+class Price extends Field
+{       
     /**
      * @inheritDoc
      */
     public static function getType(): string
     {
-        return 'post-date';
+        return 'price';
     }
 
     /**
@@ -30,7 +24,7 @@ class PostDate extends Field
      */
     public static function shouldExistOnLayout(LayoutInterface $layout): bool
     {
-        return $layout instanceof EntryLayout or $layout instanceof ProductLayout;
+        return $layout instanceof VariantLayout;
     }
 
     /**
@@ -38,7 +32,7 @@ class PostDate extends Field
      */
     public function getHandle(): string
     {
-        return 'postDate';
+        return 'price';
     }
 
     /**
@@ -46,6 +40,6 @@ class PostDate extends Field
      */
     public function getName(): string
     {
-        return \Craft::t('themes', 'Date posted');
+        return \Craft::t('themes', 'Price');
     }
 }