+++ vue/src/display/components/DisplayMenu.vue
@@ -2,8 +2,13 @@
     <nav id="notification-nav">
         <ul>
             <li class="heading"><span>{{ t('Layouts') }}</span></li>
-            <li v-for="layout2 in layouts" v-bind:key="layout2.id">
+            <li v-for="layout2 in rootLayouts" v-bind:key="layout2.id">
                 <a :href="getUrl(layout2)" :class="{'sel': layout.id === layout2.id}" @click.prevent="confirmAndChangeLayout(layout2.id)">{{ layout2.description }}</a>
+                <ul v-if="getChildren(layout2)">
+                    <li v-for="layout3 in getChildren(layout2)" v-bind:key="layout3.id">
+                        <a :href="getUrl(layout3)" :class="{'sel': layout.id === layout3.id}" @click.prevent="confirmAndChangeLayout(layout3.id)">{{ layout3.description }}</a>
+                    </li>
+                </ul>
             </li>
         </ul>
     </nav>
@@ -14,9 +19,15 @@ import { mapState, mapActions } from 'vuex';
 
 export default {
     computed: {
+        rootLayouts() {
+            return this.layouts.filter((layout) => layout.parent_id == null);
+        },
         ...mapState(['layout', 'layouts', 'theme'])
     },
     methods: {
+        getChildren(layout) {
+            return this.layouts.filter((layout2) => layout2.parent_id == layout.id);
+        },
         getUrl(layout) {
             return Craft.getCpUrl('themes/display/' + this.theme + '/' + layout.id);
         },
@@ -32,6 +43,9 @@ export default {
 </script>
 <style lang="scss" scoped>
     .heading {
-    margin: 11px 24px 14px 24px;
+        margin: 11px 24px 14px 24px;
+    }
+    .sidebar nav li ul {
+        display: block;
     }
 </style>